refactor(test): move verify_access_token tests to test_verify_access_token.py
